Add icons, update links screen to make donate button work
Previously, the static link would expire after some time. This approach uses the webview component with React Native, but the downside is that I don’t know how to get the webview to *not* update to the donation form (which doesn’t fit) when the click happens. We’d like for the webview to just show the donate button, and the form to be filled out in chrome/web browser.